+++ config/dependencies.php
@@ -312,7 +312,15 @@ return [
     \App\Controller\Msp\MspOutputController::class => function (ContainerInterface $c) {
         return new \App\Controller\Msp\MspOutputController(
             $c->get(\App\Service\LogService::class),
-            $c->get(MspOutputRepository::class)
+            $c->get(MspOutputRepository::class),
+            $c->get(TemplateRenderer::class)
+        );
+    },
+
+    \App\Controller\Msp\MspDataController::class => function (ContainerInterface $c) {
+        return new \App\Controller\Msp\MspDataController(
+            $c->get(TemplateRenderer::class),
+            $c->get(MspSensorRepository::class)
         );
     },
 
@@ -326,7 +334,15 @@ return [
     \App\Controller\N3pp\N3ppOutputController::class => function (ContainerInterface $c) {
         return new \App\Controller\N3pp\N3ppOutputController(
             $c->get(\App\Service\LogService::class),
-            $c->get(N3ppOutputRepository::class)
+            $c->get(N3ppOutputRepository::class),
+            $c->get(TemplateRenderer::class)
+        );
+    },
+
+    \App\Controller\N3pp\N3ppDataController::class => function (ContainerInterface $c) {
+        return new \App\Controller\N3pp\N3ppDataController(
+            $c->get(TemplateRenderer::class),
+            $c->get(N3ppSensorRepository::class)
         );
     },